Convert input to controlled component
Note that this is an intermediate step. Currently, the text input is a
controlled component since it has a 'change' property. The value is
being managed by state. Because of that, we have to provide an onChange
handler or the input will not be editable by the user. Read more at:

https://reactjs.org/docs/forms.html